requirements.txt
Flask==2.0.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
python-dotenv==1.0.0
Flask-Limiter==2.8.1
Flask-WTF==1.1.1
Flask-Mail==0.9.1
Authlib==1.2.0
itsdangerous==2.1.2
Werkzeug==2.1.2
bcrypt==4.0.1
setuptools==59.6.0
